requirements.txt
insightface
onnxruntime
opencv-python
pillow
matplotlib
gfpgan
gradio
huggingface_hub